UPSTREAM: ASoC: cs35l35: Correct errata handling
Currently the check of errata_chk will always evaluate to false since
the values tested don't come under the mask used. A shift of the field
is missing, add this. Also there is an error in the values tested, they
don't match the comment and the value 0x3 is not a valid value for the
field in question. Update the value to match the comment.

Signed-off-by: Charles Keepax <ckeepax@opensource.cirrus.com>
Link: https://lore.kernel.org/r/20210510131357.17170-7-ckeepax@opensource.cirrus.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 1a46b7b82df57b6b6a4e891cdbb2de1cf818a43b)

BUG=b:238403794
TEST=Test Audio use cases.

Signed-off-by: SHANKAR BOYANE <shankarx.boyane@intel.com>
Change-Id: I3131be0f941d4fc7a0b08958ff3e173025453c81